WIP: backend-frontend code integration for new branch
