+++ backend/apps/user/admin.py
@@ -1,9 +1,9 @@
 from django.contrib import admin
 
-from .models import Users
+from .models import User
 
 
 # Register your models here.
-@admin.register(Users)
+@admin.register(User)
 class UsersAdmin(admin.ModelAdmin):
     list_display = ("email", "name")
