+++ backend/apps/user/serializers.py
@@ -1,9 +1,9 @@
 from rest_framework import serializers
 
-from .models import Users
+from .models import User
 
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
-        model = Users
+        model = User
         fields = ("id", "email", "name")
